requirements.txt
requests>=2.31.0
psycopg2-binary>=2.9.9
flask>=3.0.0
python-telegram-bot>=20.7
schedule>=1.2.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
beautifulsoup4>=4.12.2
lxml>=4.9.3
fake-useragent>=1.4.0
pytz>=2023.3
